requirements.txt
aiohttp
beautifulsoup4
sqlalchemy
asyncpg
python-dotenv
apscheduler
pydantic